moved comment related code to *Comment classes, added totalComments, createdAt and modifiedAt to commentGetDto
