+++ backend/post-service/src/main/java/com/speakapp/postservice/dtos/CommentPageGetDTO.java
@@ -17,4 +17,6 @@ public class CommentPageGetDTO {
     int pageSize;
 
     int totalPages;
+
+    Long totalComments;
 }
